enable swap moves into opponents home, and define reward for non-terminal games

+++ alpha_cc/engine/__init__.pyi
@@ -48,7 +48,7 @@ class BoardInfo:
     game_over: bool
     winner: int
     size: int
-    reward: int
+    reward: float
     duration: int
 
     def __init__(self, size: int) -> None: ...
